refactor: moved japrotocol to nw module
