+++ jarust/examples/secure_ws.rs
@@ -1,9 +1,9 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust::japlugin::Attach;
-use jarust::japrotocol::EstablishmentProtocol;
-use jarust::japrotocol::Jsep;
-use jarust::japrotocol::JsepType;
+use jarust::nw::japrotocol::EstablishmentProtocol;
+use jarust::nw::japrotocol::Jsep;
+use jarust::nw::japrotocol::JsepType;
 use jarust::params::AttachHandleParams;
 use jarust::params::CreateConnectionParams;
 use jarust::TransactionGenerationStrategy;
