+++ jarust/src/prelude.rs
@@ -3,7 +3,7 @@ pub use crate::jaconnection::JaResponseStream;
 pub use crate::jahandle::JaHandle;
 pub use crate::japlugin::Attach;
 pub use crate::japlugin::PluginTask;
-pub use crate::japrotocol::JaResponse;
 pub use crate::jasession::JaSession;
+pub use crate::nw::japrotocol::JaResponse;
 
 pub type JaResult<T> = core::result::Result<T, JaError>;
