+++ jarust/tests/mocks/mock_handle.rs
@@ -1,7 +1,7 @@
 use super::mock_transport::MockServer;
-use jarust::japrotocol::JaData;
-use jarust::japrotocol::JaSuccessProtocol;
-use jarust::japrotocol::ResponseType;
+use jarust::nw::japrotocol::JaData;
+use jarust::nw::japrotocol::JaSuccessProtocol;
+use jarust::nw::japrotocol::ResponseType;
 use jarust::params::AttachHandleParams;
 use jarust::prelude::*;
 use std::time::Duration;
