+++ jarust_plugins/src/echo_test/events.rs
@@ -1,9 +1,9 @@
 use jarust::error::JaError;
-use jarust::japrotocol::EstablishmentProtocol;
-use jarust::japrotocol::GenericEvent;
-use jarust::japrotocol::JaHandleEvent;
-use jarust::japrotocol::JaResponse;
-use jarust::japrotocol::ResponseType;
+use jarust::nw::japrotocol::EstablishmentProtocol;
+use jarust::nw::japrotocol::GenericEvent;
+use jarust::nw::japrotocol::JaHandleEvent;
+use jarust::nw::japrotocol::JaResponse;
+use jarust::nw::japrotocol::ResponseType;
 use serde::Deserialize;
 use serde_json::from_value;
 
@@ -66,13 +66,13 @@ impl TryFrom<JaResponse> for PluginEvent {
 mod tests {
     use super::PluginEvent;
     use crate::echo_test::events::EchoTestEvent;
-    use jarust::japrotocol::EstablishmentProtocol;
-    use jarust::japrotocol::JaHandleEvent;
-    use jarust::japrotocol::JaResponse;
-    use jarust::japrotocol::Jsep;
-    use jarust::japrotocol::JsepType;
-    use jarust::japrotocol::PluginData;
-    use jarust::japrotocol::ResponseType;
+    use jarust::nw::japrotocol::EstablishmentProtocol;
+    use jarust::nw::japrotocol::JaHandleEvent;
+    use jarust::nw::japrotocol::JaResponse;
+    use jarust::nw::japrotocol::Jsep;
+    use jarust::nw::japrotocol::JsepType;
+    use jarust::nw::japrotocol::PluginData;
+    use jarust::nw::japrotocol::ResponseType;
     use serde_json::json;
 
     #[test]
