+++ jarust_plugins/src/video_room/events.rs
@@ -1,6 +1,6 @@
 use jarust::error::JaError;
-use jarust::japrotocol::GenericEvent;
-use jarust::japrotocol::JaResponse;
+use jarust::nw::japrotocol::GenericEvent;
+use jarust::nw::japrotocol::JaResponse;
 
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 pub enum PluginEvent {
